Add Gateway object to interact with Open North API
Start interacting with the Open North API by retrieving the list of
candidates for a given postal code. We then want to filter the
candidates for only the 4 major parties (lots of little parties we don't
care about). Then return only the 3 pieces of information we want:
- name
- party
- riding

All this is then returned in a callback since this is an API call.